+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraKeyValueServiceImpl.java
@@ -1757,10 +1757,23 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     @Override
     public void deleteAllTimestamps(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell,
             boolean deleteSentinels) {
+        if (maxTimestampExclusiveByCell.isEmpty()) {
+            return;
+        }
+
         Map<InetSocketAddress, Map<Cell, Long>> keysByHost = HostPartitioner.partitionMapByHost(
                 clientPool, maxTimestampExclusiveByCell.entrySet());
+
+        long maxTimestampForAllCells = maxTimestampExclusiveByCell.values().stream().mapToLong(x -> x).max().getAsLong();
+        long rangeTombstoneCassandraTimestamp =
+                mutationTimestampProvider.getRangeTombstoneTimestamp(maxTimestampForAllCells);
         for (Map.Entry<InetSocketAddress, Map<Cell, Long>> entry : keysByHost.entrySet()) {
-            deleteAllTimestampsOnSingleHost(tableRef, entry.getKey(), entry.getValue(), deleteSentinels);
+            deleteAllTimestampsOnSingleHost(
+                    tableRef,
+                    entry.getKey(),
+                    entry.getValue(),
+                    deleteSentinels,
+                    rangeTombstoneCassandraTimestamp);
         }
     }
 
@@ -1768,7 +1781,8 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
             TableReference tableRef,
             InetSocketAddress host,
             Map<Cell, Long> maxTimestampExclusiveByCell,
-            boolean deleteSentinels) {
+            boolean deleteSentinels,
+            long rangeTombstoneCassandraTimestamp) {
         if (maxTimestampExclusiveByCell.isEmpty()) {
             return;
         }
@@ -1797,11 +1811,11 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     }
 
     private void insertRangeTombstones(CassandraClient client, Map<Cell, Long> maxTimestampExclusiveByCell,
-            TableReference tableRef, boolean deleteSentinel) throws TException {
+            TableReference tableRef, boolean deleteSentinel, long rangeTombstoneCassandraTs) throws TException {
         MutationMap mutationMap = new MutationMap();
 
         maxTimestampExclusiveByCell.forEach((cell, maxTimestampExclusive) -> {
-            Mutation mutation = getMutation(cell, maxTimestampExclusive, deleteSentinel);
+            Mutation mutation = getMutation(cell, maxTimestampExclusive, deleteSentinel, rangeTombstoneCassandraTs);
 
             mutationMap.addMutationForCell(cell, tableRef, mutation);
         });
@@ -1810,15 +1824,16 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
                 deleteConsistency);
     }
 
-    private Mutation getMutation(Cell cell, long maxTimestampExclusive, boolean deleteSentinel) {
+    private Mutation getMutation(Cell cell, long maxTimestampExclusive,
+            boolean deleteSentinel, long rangeTombstoneCassandraTimestamp) {
         if (deleteSentinel) {
             return Mutations.rangeTombstoneIncludingSentinelForColumn(cell.getColumnName(), maxTimestampExclusive,
-                    mutationTimestampProvider.getRangeTombstoneTimestamp(maxTimestampExclusive));
+                    rangeTombstoneCassandraTimestamp);
         }
         return Mutations.rangeTombstoneForColumn(
                 cell.getColumnName(),
                 maxTimestampExclusive,
-                mutationTimestampProvider.getRangeTombstoneTimestamp(maxTimestampExclusive));
+                rangeTombstoneCassandraTimestamp);
     }
 
     /**